move Sampler to breeze channel
