+++ vNext/channels/applicationinsights-channel-js/Tests/Selenium/aichannel.tests.ts
@@ -1,5 +1,7 @@
 ﻿import { SenderTests } from "../Sender.tests";
+import { SampleTests } from "../Sample.tests";
 
 export function runTests() {
     new SenderTests().registerTests();
+    new SampleTests().registerTests();
 }
