+++ vNext/channels/applicationinsights-channel-js/src/Interfaces.ts
@@ -43,6 +43,8 @@ export interface ISenderConfig {
     instrumentationKey: () => string;
 
     namePrefix: () => string;
+
+    sampleRate: () => number;
 }
 
 export interface IBackendResponse {
